feat: make `collectSourceAttributions` available
